spotnet/pyproject.toml
[tool.poetry]
name = "spotnet"
version = "0.1.0"
description = "The functionality allows users to leverage their positions on assets (e.g., ETH) by looping through lending protocols (ZkLend) and automated market makers (AMMs)."
authors = ["djeck1432 <jack_horobets@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "0.115.0"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.1"
uvicorn = "0.30.6"
starknet-py = "0.23.0"
jinja2 = "3.1.4"
python-multipart = "0.0.9"
itsdangerous = "2.2.0"
alembic = "1.13.3"
sqlalchemy = "2.0.35"
pytest = "8.3.3"
pytest-cov = "6.0.0"
pytest-asyncio = "0.24.0"
pytest-env = "1.1.5"
pytest-mock = "3.14.0"
httpx = "0.27.2"
celery = "5.4.0"
redis = "5.2.0"
trio = "0.27.0"
aiogram = ">=3.13.1"
aiohttp = "^3.10.10"
psycopg2 = "^2.9.10"
faker = "^30.8.1"
notebook = "^7.2.2"
sentry-sdk = {extras = ["fastapi"], version = "^2.18.0"}
pragma-sdk = "^2.4.6"

[tool.poetry.group.dev.dependencies]
black = "24.8.0"
isort = "5.13.2"
pre-commit = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
